+++ identity_server/logic/validation_chain/handlers/header_validation_handler.py
@@ -1,5 +1,3 @@
-from typing import Optional
-
 from identity_server.logic.validation_chain.handler import Handler
 from identity_server.logic.validation_chain.exceptions.validator_exceptions import HeaderValidationException
